# docs/swagger.yaml
basePath: /v1
definitions:
...
  entity.CreateProductRequestForSwagger:
    properties:
      business_id:
        type: string
      category_id:
        type: string
      cost:
        type: integer
      count:
        type: integer
      description:
        type: string
      discount:
        type: integer
      discount_cost:
        type: integer
      name:
        type: string
      short_info:
        type: string
    type: object
...
  entity.GetAllProductsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/entity.Product'
        type: array
      total:
        type: integer
    type: object
...
  entity.Product:
    properties:
      business_id:
        type: string
      category_id:
        type: string
      cost:
        type: integer
      count:
        type: integer
      created_at:
        type: string
      description:
        type: string
      discount:
        type: integer
      discount_cost:
        type: integer
      id:
        type: string
      name:
        type: string
      short_info:
        type: string
      updated_at:
        type: string
    type: object
...
  entity.UpdateProductRequest:
    properties:
      category_id:
        type: string
      cost:
        type: integer
      count:
        type: integer
      description:
        type: string
      discount:
        type: integer
      discount_cost:
        type: integer
      id:
        type: string
      name:
        type: string
      short_info:
        type: string
    type: object
...
  status_http.Response:
    properties:
      custom_message: {}
      data: {}
      description:
        type: string
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Using a translation service as an example
  title: Go Clean Template API
  version: "1.0"
paths:
...
  /product/create:
    post:
      consumes:
      - application/json
      description: Create a new product in the database
      parameters:
      - description: Product Details
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/entity.CreateProductRequestForSwagger'
      produces:
      - application/json
      responses:
        "201":
          description: Product created successfully
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Create a new product
      tags:
      - PRODUCT
  /product/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a product from the database by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Product deleted successfully
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad request
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Delete a product by ID
      tags:
      - PRODUCT
  /product/get/{id}:
    get:
      consumes:
      - application/json
      description: Get product details by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Product data
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Product Not Found
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get product by ID
      tags:
      - PRODUCT
  /product/list:
    get:
      consumes:
      - application/json
      description: Get a paginated list of products with optional filters
      parameters:
      - description: Filter by Category ID
        in: query
        name: category_id
        type: string
      - description: Search in name, description, or short_info
        in: query
        name: search
        type: string
      - description: 'Number of products per page (default: 10)'
        in: query
        name: limit
        required: true
        type: integer
      - description: 'Page number (starts from 1, default: 1)'
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of Products
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/entity.GetAllProductsResponse'
              type: object
        "400":
          description: Bad request
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: List products
      tags:
      - PRODUCT
  /product/update:
    put:
      consumes:
      - application/json
      description: Update product details by ID
      parameters:
      - description: Product Details
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/entity.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Update an existing product
      tags:
      - PRODUCT
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"